promote: checking msg.guild when exiting

+++ Commands/Public/promote.js
@@ -54,8 +54,10 @@ module.exports = async(client, msg, suffix) => {
 	//
 	// 	if (msg.guild && collected) collected.delete().catch(e => null);
 	// 	if (!collected || /^0$/.test(collected.content)) {
+	//		if (msg.guild) {
 	// 		pmsg.delete().catch(e => null);
 	// 		omsg.delete().catch(e => null);
+	//		}
 	// 		msg.author.busy = false;
 	//		return;
 	// 	}
@@ -86,8 +88,10 @@ module.exports = async(client, msg, suffix) => {
 
 		if (msg.guild && collected) collected.delete().catch(e => null);
 		if (!collected || /^0$/.test(collected.content)) {
-			pmsg.delete().catch(e => null);
-			omsg.delete().catch(e => null);
+			if (msg.guild) {
+				pmsg.delete().catch(e => null);
+				omsg.delete().catch(e => null);
+			}
 			msg.author.busy = false;
 			return;
 		}
@@ -118,8 +122,10 @@ module.exports = async(client, msg, suffix) => {
 
 		if (msg.guild && collected) collected.delete().catch(e => null);
 		if (!collected || /^0$/.test(collected.content) || (!setup && /^skip$/i.test(collected.content))) {
-			pmsg.delete().catch(e => null);
-			omsg.delete().catch(e => null);
+			if (msg.guild) {
+				pmsg.delete().catch(e => null);
+				omsg.delete().catch(e => null);
+			}
 			msg.author.busy = false;
 			return;
 		}
@@ -160,8 +166,10 @@ module.exports = async(client, msg, suffix) => {
 			)).first();
 
 			if (!collected || /^0$/.test(collected.content)) {
-				pmsg.delete().catch(e => null);
-				omsg.delete().catch(e => null);
+				if (msg.guild) {
+					pmsg.delete().catch(e => null);
+					omsg.delete().catch(e => null);
+				}
 				msg.author.busy = false;
 				return;
 			}
@@ -188,7 +196,9 @@ module.exports = async(client, msg, suffix) => {
 
 			if (msg.guild && collected) collected.delete().catch(e => null);
 			if (!collected || /^0$/.test(collected.content)) {
-				omsg.delete().catch(e => null);
+				if (msg.guild) {
+					omsg.delete().catch(e => null);
+				}
 				msg.author.busy = false;
 				return;
 			}
@@ -211,8 +221,10 @@ module.exports = async(client, msg, suffix) => {
 
 			if (msg.guild && collected) collected.delete().catch(e => null);
 			if (!collected || /^0$/.test(collected.content)) {
-				pmsg.delete().catch(e => null);
-				omsg.delete().catch(e => null);
+				if (msg.guild) {
+					pmsg.delete().catch(e => null);
+					omsg.delete().catch(e => null);
+				}
 				msg.author.busy = false;
 				return;
 			}
@@ -330,7 +342,9 @@ module.exports = async(client, msg, suffix) => {
 
 	if (msg.guild && collected) collected.delete().catch(e => null);
 	if (!collected || /^0$/.test(collected.content)) {
-		omsg.delete().catch(e => null);
+		if (msg.guild) {
+			omsg.delete().catch(e => null);
+		}
 		msg.author.busy = false;
 		return;
 	}
@@ -377,8 +391,10 @@ module.exports = async(client, msg, suffix) => {
 		}
 
 		case "3": {
-			pmsg.delete().catch(e => null);
-			omsg.delete().catch(e => null);
+			if (msg.guild) {
+				pmsg.delete().catch(e => null);
+				omsg.delete().catch(e => null);
+			}
 			msg.author.busy = false;
 			return msg.channel.send({ embed: { color: config.colors.error, title: "Unfinished feature", description: "Please use (1) for now." } });
 		}
